requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pillow==11.3.0
